pyproject.toml
[project]
name = "visa-center"
version = "0.1.0"
description = ""
authors = [
    {name = "begench2005",email = "begenchallakov070@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
